+++ Twilio/Rest/Video/V1/RecordingInstance.php
@@ -28,6 +28,7 @@ use Twilio\Version;
  * @property string containerFormat
  * @property string codec
  * @property array groupingSids
+ * @property string trackName
  * @property array links
  */
 class RecordingInstance extends InstanceResource {
@@ -56,6 +57,7 @@ class RecordingInstance extends InstanceResource {
             'containerFormat' => Values::array_get($payload, 'container_format'),
             'codec' => Values::array_get($payload, 'codec'),
             'groupingSids' => Values::array_get($payload, 'grouping_sids'),
+            'trackName' => Values::array_get($payload, 'track_name'),
             'links' => Values::array_get($payload, 'links'),
         );