+++ Twilio/Rest/Video/V1/Room/ParticipantContext.php
@@ -13,15 +13,18 @@ use Twilio\Exceptions\TwilioException;
 use Twilio\InstanceContext;
 use Twilio\Options;
 use Twilio\Rest\Video\V1\Room\Participant\PublishedTrackList;
+use Twilio\Rest\Video\V1\Room\Participant\SubscribedTrackList;
 use Twilio\Values;
 use Twilio\Version;
 
 /**
  * @property \Twilio\Rest\Video\V1\Room\Participant\PublishedTrackList publishedTracks
+ * @property \Twilio\Rest\Video\V1\Room\Participant\SubscribedTrackList subscribedTracks
  * @method \Twilio\Rest\Video\V1\Room\Participant\PublishedTrackContext publishedTracks(string $sid)
  */
 class ParticipantContext extends InstanceContext {
     protected $_publishedTracks = null;
+    protected $_subscribedTracks = null;
 
     /**
      * Initialize the ParticipantContext
@@ -103,6 +106,23 @@ class ParticipantContext extends InstanceContext {
         }
 
         return $this->_publishedTracks;
+    }
+
+    /**
+     * Access the subscribedTracks
+     * 
+     * @return \Twilio\Rest\Video\V1\Room\Participant\SubscribedTrackList 
+     */
+    protected function getSubscribedTracks() {
+        if (!$this->_subscribedTracks) {
+            $this->_subscribedTracks = new SubscribedTrackList(
+                $this->version,
+                $this->solution['roomSid'],
+                $this->solution['sid']
+            );
+        }
+
+        return $this->_subscribedTracks;
     }
 
     /**