+++ CHANGES.md
@@ -1,6 +1,24 @@
 twilio-php Changelog
 ====================
 
+[2018-01-30] Version 5.16.4
+----------------------------
+**Api**
+- Add `studio-engagements` usage key
+
+**Preview**
+- Remove Studio Engagement Deletion
+
+**Studio**
+- Initial Release
+
+**Video**
+- [omit] Beta: Allow updates to `SubscribedTracks`.
+- Add `SubscribedTracks`.
+- Add track name to Video Recording resource
+- Add Composition and Composition Media resources
+
+
 [2018-01-19] Version 5.16.3
 ----------------------------
 **Api**